requirements.txt
alembic==1.12.0
Authlib==1.2.1
azure-common==1.1.28
azure-core==1.29.4
azure-identity==1.13.0
azure-mgmt-core==1.4.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.0.1
azure-mgmt-subscription==3.1.1
Babel==2.13.0
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.3
boto3==1.28.65
botocore==1.31.65
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
dnspython==2.4.2
email-validator==2.0.0.post2
eventlet==0.33.3
Flask==2.2.5
flask-babel==3.1.0
Flask-Compress==1.14
Flask-Gravatar==0.5.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.1.2
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
flask-swagger-ui==4.11.1
Flask-WTF==1.1.1
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.61.0
greenlet==1.1.2
h11==0.14.0
httpagentparser==1.9.5
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
keyring==23.13.1
ldap3==2.9.1
Mako==1.2.4
MarkupSafe==2.1.3
more-itertools==10.1.0
msal==1.24.1
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
paramiko==3.3.1
passlib==1.7.4
pgadmin4==7.7
Pillow==9.5.0
portalocker==2.8.2
protobuf==4.24.4
psutil==5.9.6
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.7.0
SecretStorage==3.3.3
simple-websocket==1.0.0
six==1.16.0
speaklater3==1.4
SQLAlchemy==2.0.22
sqlparse==0.4.4
sshtunnel==0.4.0
typing_extensions==4.8.0
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
Werkzeug==2.2.3
wsproto==1.2.0
WTForms==3.0.1
zipp==3.17.0
